Finished busy wait
Also found an int overflow bug that I fixed using
longs instead. Fun times!

+++ CentralComputing/Utils.h
@@ -26,7 +26,7 @@ extern LogLevel loglevel;
 
 void print(LogLevel level, const char * format, ...);
 
-void busyWait(int microseconds);
+void busyWait(long microseconds);
 
 /**
 * Prints errno, with user defined message prepended
